+++ src/window_type.h
@@ -695,7 +695,7 @@ enum GameOptionsInvalidationData {
 struct Window;
 
 /** Number to differentiate different windows of the same class */
-typedef int32 WindowNumber;
+typedef int64 WindowNumber;
 
 /** State of handling an event. */
 enum EventState {
